requirements.txt
# Core ML — install torch with CUDA on pod via setup_pod.sh; CPU-only here for local dev
torch>=2.2.0
transformers>=4.40.0
peft>=0.10.0
bitsandbytes>=0.43.0
accelerate>=0.28.0
datasets>=2.18.0

# Data / config
pydantic>=2.0.0
python-dotenv>=1.0.0

# OpenAI API
openai>=1.20.0

# Utilities
numpy>=1.26.0
scipy>=1.12.0
tqdm>=4.66.0

# Analysis / plotting (optional on pod)
pandas>=2.1.0
matplotlib>=3.8.0
